src/requirements.txt
pillow>=12.1.1
reportlab>=3.6.0